+++ scooterMS/src/main/java/tpe/scooterMS/repository/StopRepository.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
 import tpe.scooterMS.model.Stop;
 
 @Repository("stopRepository")
-public interface StopRepository extends JpaRepository<Stop, Integer>{
+public interface StopRepository extends JpaRepository<Stop, Long>{
 	@Query("SELECT s FROM Stop s WHERE s.id = :id")
 	public Stop getStopById(@Param("id")long id);
